patch: increase timeout
By default, mocha will time out after 2 seconds. This may be too
short of a run, as it is latency-dependent. This change aims to
increase the timeout to a reasonable 30 seconds: long enough for
the test to accomplish what it's supposed to and short enough that
it should timeout when there are other factors involved